refactor(player): move generic code to sdk
