+++ sdk/src/player/sc/player/PlayerClient.kt
@@ -1,7 +1,7 @@
-package sc.plugin2021
+package sc.player
 
 import org.slf4j.LoggerFactory
-import sc.api.plugins.IGameState
+import sc.api.plugins.*
 import sc.framework.plugins.protocol.MoveRequest
 import sc.networking.clients.AbstractLobbyClientListener
 import sc.networking.clients.IControllableGame
@@ -13,6 +13,7 @@ import sc.protocol.responses.ProtocolMessage
 import sc.shared.GameResult
 import sc.shared.WelcomeMessage
 import java.net.ConnectException
+import java.util.ServiceLoader
 import kotlin.system.exitProcess
 
 /**
@@ -22,7 +23,7 @@ import kotlin.system.exitProcess
 class PlayerClient(
         host: String,
         port: Int,
-        private val handler: IGameHandler
+        private val handler: IGameHandler,
 ): AbstractLobbyClientListener() {
     companion object {
         private val logger = LoggerFactory.getLogger(PlayerClient::class.java)
@@ -46,24 +47,20 @@ class PlayerClient(
     private lateinit var roomId: String
     
     /** The team the client belongs to. Needed to connect client and player. */
-    var team: Team? = null
+    var teamName: String? = null
         private set
-    
-    /** Tell this client to observe the game given by the preparation handler. */
-    fun observeGame(handle: GamePreparedResponse): IControllableGame =
-            client.observe(handle.roomId)
     
     /** Called for any new message sent to the game room, e.g., move requests. */
     override fun onRoomMessage(roomId: String, data: ProtocolMessage) {
         this.roomId = roomId
-        when(data) {
+        when (data) {
             is MoveRequest -> sendMove(handler.calculateMove())
-            is WelcomeMessage -> team = Team.valueOf(data.color.toUpperCase())
+            is WelcomeMessage -> teamName = data.color
         }
     }
     
     /** Sends the selected move to the server. */
-    fun sendMove(move: Move) =
+    fun sendMove(move: IMove) =
             client.sendMessageToRoom(roomId, move)
     
     /** Called when an erroneous message is sent to the room. */
@@ -77,18 +74,17 @@ class PlayerClient(
      * Happens after a client made a move.
      */
     override fun onNewState(roomId: String, state: IGameState) {
-        val gameState = state as GameState
+        val gameState = state as TwoPlayerGameState<*>
         logger.debug("$this got a new state $gameState")
-    
-        if (team == null || !gameState.hasValidColors())
-            return
-    
-        if (gameState.currentTeam == team) {
-            handler.onUpdate(gameState.currentPlayer, gameState.otherPlayer)
-        } else {
-            handler.onUpdate(gameState.otherPlayer, gameState.currentPlayer)
+        
+        teamName?.let { teamName ->
+            if (gameState.currentTeam.name == teamName) {
+                handler.onUpdate(gameState.currentPlayer, gameState.otherPlayer)
+            } else {
+                handler.onUpdate(gameState.otherPlayer, gameState.currentPlayer)
+            }
+            handler.onUpdate(gameState)
         }
-        handler.onUpdate(gameState)
     }
     
     /** Start the LobbyClient [client] and listen to it. */
@@ -100,7 +96,7 @@ class PlayerClient(
     /** [start] and join any game with the appropriate [gameType]. */
     fun joinAnyGame() {
         start()
-        client.joinRoomRequest(GamePlugin.PLUGIN_ID)
+        client.joinRoomRequest(IGamePlugin.loadPluginId())
     }
     
     override fun onGameLeft(roomId: String) {
@@ -111,7 +107,7 @@ class PlayerClient(
     override fun onGameOver(roomId: String, data: GameResult) {
         logger.info("$this: Game over with result $data")
         isGameOver = true
-        handler.gameEnded(data, team, error)
+        handler.gameEnded(data, error)
     }
     
     fun joinPreparedGame(reservation: String) {
